+++ src/features/admin/admin.routes.ts
@@ -1,6 +1,9 @@
 import { Router } from "express";
 import {
   approveSession,
+  rejectSession,
+  showLivePostDetail,
+  showLivePosts,
   showPendingPosts,
   showSessionPosts,
 } from "./admin.controller.js";
@@ -13,6 +16,10 @@ router.get("/admin/queue/:id", showSessionPosts);
 
 router.post("/admin/queue/:id/approve", approveSession);
 
-// router.post("/admin/queue/:id/reject", rejectSession);
+router.post("/admin/queue/:id/reject", rejectSession);
+
+router.get("/admin/liveposts", showLivePosts);
+
+router.get("/admin/liveposts/:id", showLivePostDetail);
 
 export default router;