+++ src/shared/error.ts
@@ -2,7 +2,11 @@ export type appError =
   | { reason: "NO_SESSION" }
   | { reason: "DB_ERROR" }
   | { reason: "SESSION_NOT_FOUND" }
-  | { reason: "JOBS_NOT_FOUND" };
+  | { reason: "POSTS_NOT_FOUND" }
+  | { reason: "POST_NOT_FOUND" }
+  | { reason: "EMAIL_API_ERROR" }
+  | { reason: "TOKEN_NOT_FOUND" }
+  | { reason: "TOKEN_EXPIRED" };
 
 export type Result<T> =
   | { success: true; data: T }